Allow to extra configuration for mock if needed
Phpunit is a complex library and sometimes you need
to configure a mock in a particular way.

Even the intent of this library is to not create
another mocking tool, allowing custom mock configuration
is a must in this case.

So the `mock` method got a new optional parameter that
get passed a `MockBuilder` instance which allow to
configure the mock before the mock is created.